DFK-354 - Oppdaterer ikke CrawlResultat.Ferdig som allerede finnes

+++ src/main/kotlin/no/uutilsynet/testlab2testing/maaling/MaalingDAO.kt
@@ -607,6 +607,19 @@ class MaalingDAO(val jdbcTemplate: NamedParameterJdbcTemplate) {
 
   @Transactional
   fun saveCrawlResultat(crawlResultat: CrawlResultat, maalingId: Int) {
+    val alreadyFinished =
+        jdbcTemplate.queryForObject(
+            "select count(*) from crawlresultat where loeysingid = :loeysingid and maaling_id = :maaling_id and status = :status_finished",
+            mapOf(
+                "loeysingid" to crawlResultat.loeysing.id,
+                "maaling_id" to maalingId,
+                "status_finished" to "ferdig"),
+            Int::class.java)
+
+    if (alreadyFinished == 1) {
+      return
+    }
+
     jdbcTemplate.update(
         "delete from crawlresultat where loeysingid = :loeysingid and maaling_id = :maaling_id",
         mapOf("loeysingid" to crawlResultat.loeysing.id, "maaling_id" to maalingId))
